requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.8.2
python-multipart==0.0.5
python-dotenv==1.0.0
email-validator==2.1.0
fastapi-mail==0.3.0
gunicorn==20.1.0